refactor: Move Board and Player classes into their own modules
- Move Board and Player classes into their own modules and import them in main module
- Add function to introduce game and its rules (feat)
